+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
-## [Unreleased]
+
+## [20201026]
+
+  - Configurable allow origins headers
+  - Add unit tests
+  - Handle The Graph errors
+  - Use the external pools package, refs #1
+  - Setup CI, refs #2
+  - Document response struct, refs #4
+  - Setup Sentry, refs #5
+  - New tokens daily endpoint, refs #6
+
+
+## [20201014]
 
   - Initial release